* Added createArray and destroyArray to memory::MemoryWorker * Added arithmetic operators to Pointer class
git-svn-id: http://subversion.assembla.com/svn/kraft@34 06128d7d-dce7-46cd-bccc-bdfe512662dc

+++ projects/kq/code/source/kqKlarity.cpp
@@ -181,7 +181,7 @@ public:
 		}
 
 		if(m_fLog){
-			fwrite(m_sBuffer, nChars, 1, m_fLog);
+			fwrite(m_sBuffer, (size_t)nChars, 1, m_fLog);
 			fflush(m_fLog);
 		}
